requirements.txt
bs4
filelock
importlib_metadata>=4.2.0
jieba
packaging
regex
sacremoses
scikit-learn
scipy==1.5.4
sentencepiece>=0.1.91
tokenizers==0.10.1
datasets==2.1.0
rouge
tqdm>=4.48.2
pillow
albumentations
einops
